requirements.txt
aiohttp
beautifulsoup4
discord.py
google-api-python-client
lxml
pafy
Pillow
psutil
xmltodict
youtube-dl
